# docs/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: hajaCheck API
  version: "0.38.0-draft"
  description: >
    Contract-First 초안 (PRD §6, §8). 실제 소스 오브 트루스는 이 파일.


    ## 실시간(WS) 상담 — /ws STOMP (FR-7, #20/HAJA-33, #1002/HAJA-496)

    OpenAPI 3.0 은 WebSocket/STOMP 를 네이티브로 표현하지 못하므로, 이 계약은 `paths` 항목이 아니라
    이 서술로 대체한다(`backend/.../counsel/websocket/WebSocketConfig.java` 실구현 기준).

    - **엔드포인트**: `/ws` (SockJS 미사용 — dev(Vite)/prod(nginx) 동일 오리진 프록시 전제, 기본 same-origin
      정책 유지).
    - **인증**: SESSION 쿠키 기반, 2단계로 강제한다 — ① 핸드셰이크(`CounselHandshakeInterceptor`)에서
      SESSION 쿠키 검증, ② 인바운드 채널(`StompAuthChannelInterceptor`)에서 CONNECT 재검증 + SUBSCRIBE
      시점 당사자(참여자) 검증. 클라이언트 페이로드의 senderId 는 신뢰하지 않고 핸드셰이크에서 확립한
      `StompUserPrincipal` 로만 발신자를 식별한다.
    - **destinations**:
      - `SEND /app/counsel/{ticketId}/send` — 상담 메시지 전송(`ChatMessageSendRequest`: content
        ≤2000자/attachmentKey ≤500자, 텍스트·첨부 중 최소 하나). 티켓이 IN_PROGRESS 가 아니거나 발신자가
        참여자(사용자 본인/담당 상담원)가 아니면 예외 대신 조용히 드롭한다(WS 프레임 처리 중 예외로 세션이
        끊기는 것을 피하기 위한 방어적 설계).
      - `SUBSCRIBE /topic/counsel/{ticketId}` — 해당 상담방 브로드캐스트 구독. 서버가 저장된 메시지를
        `ChatMessageResponse` 로 이 destination 에 `convertAndSend` 한다.
      - `SUBSCRIBE /user/queue/counsel/assigned` — 개인 알림(user destination prefix `/user`). 내 티켓이
        상담원에게 배정되면(`POST /api/counsel/tickets/{id}/assign` 성공 시) `CounselTicketResponse` 를
        수신한다.
      - `SUBSCRIBE /user/queue/counsel/ended` — 개인 알림. 내 티켓이 종료되면(`POST /api/counsel/tickets/
...


...


...


...


...


...


...


...


...

...

...

...
servers:
...
paths:
...
  /api/inspections:
    post:
      summary: 점검(회차) 생성
      description: >
        인증 사용자의 유효 회사가 소유한 시설에만 점검 회차를 생성하고 담당 점검자를 지정한다.
        assignedInspectorId는 같은 회사 소속이며 ACTIVE인 INSPECTOR 또는 ADMIN만 허용하며,
        관리자와 담당 점검자에게도 타회사 시설 접근 예외는 없다.
        createdBy는 요청 바디가 아니라 인증 주체에서 서버가 기록한다.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [inspection]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InspectionCreateRequest" }
      responses:
        "201":
          description: 점검 회차 생성 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InspectionApiResponse" }
        "400":
          description: 입력값 또는 담당 점검자 상태/역할 오류
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "401":
          description: 미인증
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: 시설 접근 권한 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
    get:
      summary: 점검(회차) 목록 조회 (#725/HAJA-393, PR #891로 실구현 완료·PR#1032 사람검수 P2-3로 소급 계약화)
      description: >
        인증 사용자 소유 회사(회사 스코프)의 시설물에 속한 점검 회차를 상태/유형/점검일/회차/전체
        하자 건수/하자 조건으로 필터링해 페이지 단위로 반환한다. 배열 축은 배열 내부 OR, 서로 다른
        축은 AND다. defectType/defectGrade/defectStatus는 같은 비삭제 하자 하나가 주어진 조건을
        전부 만족해야 한다. defectCount는 하자 조건과 무관하게 해당 점검의 deleted=false 전체 하자
        건수이며 응답 defectCount와 동일한 집계다.
      security:
        - SessionCookie: []
      tags: [inspection]
      parameters:
        - name: facilityId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          description: 점검 진행 상태(복수 가능, 반복 키). 기존 단일 status 요청도 동일하게 동작한다.
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/InspectionStatusCode" }
        - name: inspectionType
          in: query
          description: 점검 유형(복수 가능, 반복 키)
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/InspectionTypeCode" }
        - name: inspectionDateFrom
          in: query
          description: 점검일 시작일(포함)
          schema: { type: string, format: date }
        - name: inspectionDateTo
          in: query
          description: 점검일 종료일(포함)
          schema: { type: string, format: date }
        - name: roundNoMin
          in: query
          description: 시설물 기준 점검 회차 하한(포함)
          schema: { type: integer, minimum: 1 }
        - name: roundNoMax
          in: query
          description: 시설물 기준 점검 회차 상한(포함)
          schema: { type: integer, minimum: 1 }
        - name: defectCountMin
          in: query
          description: 비삭제 전체 하자 건수 하한(포함)
          schema: { type: integer, format: int64, minimum: 0 }
        - name: defectCountMax
          in: query
          description: 비삭제 전체 하자 건수 상한(포함)
          schema: { type: integer, format: int64, minimum: 0 }
        - name: defectType
          in: query
          description: 하자 유형 필터(복수 가능, 반복 키 defectType=A&defectType=B — 대괄호 없는 형태)
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/DefectTypeCode" }
        - name: defectGrade
          in: query
          description: 하자 등급 필터(복수 가능, 반복 키)
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/DefectGradeCode" }
        - name: defectStatus
          in: query
          description: 하자 상태 필터(복수 가능, 반복 키)
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/DefectStatusCode" }
        - name: page
          in: query
          description: 0부터 시작하는 페이지 번호
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          description: 페이지 크기
          schema: { type: integer, minimum: 1, default: 20 }
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InspectionListItemPageApiResponse" }
        "400":
          description: enum/날짜 형식/하한 또는 역전된 범위가 올바르지 않음(INVALID_INPUT)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "401":
          description: 미인증
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
...
  /api/inspections/{id}/analyze:
    post:
      summary: AI 분석 요청
      tags: [inspection]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "202": { description: TODO — 황승현 확정 필요(비동기 잡 패턴) }
...
components:
  securitySchemes:
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: Spring Boot가 FastAPI 내부 호출에 부착하는 AI_INTERNAL_KEY
    InternalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
      description: Spring Boot와 FastAPI 환경변수에만 주입하는 내부 호출 토큰
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Spring Security 인증 세션 쿠키
    XsrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: 상태 변경 요청에 필요한 double-submit CSRF 토큰
  schemas:
...
    ApiError:
      type: object
      properties:
        code: { type: string, example: AUTH_INVALID_CREDENTIALS }
        message: { type: string, example: 아이디 또는 비밀번호가 올바르지 않습니다. }
...
    ErrorApiResponse:
      type: object
      description: 실패 envelope — error=ApiError
      properties:
        success: { type: boolean, example: false }
        data: { type: object, nullable: true, example: null }
        error: { $ref: "#/components/schemas/ApiError" }

    # ── counsel 도메인 (FR-7, #20/HAJA-33, #1002/HAJA-496) ──
...
    DefectTypeCode:
      type: string
      description: >-
        탐지 클래스 3종 확정(PRD v0.42, 2026-07-13) — LEAK_EFFLORESCENCE(누수·백태)·PAINT_DAMAGE(도장손상)는
        데이터 확보 상황상 이번 범위에서 제외됨. DB `defect_type` enum·프론트 DefectType은 별도 확인·정합 필요.
      enum: [CRACK, SPALLING, REBAR_EXPOSURE]
    DefectGradeCode:
      type: string
      enum: [A, B, C, D, E]
    DefectStatusCode:
      type: string
      description: >-
        하자 조치 상태 4단계(DETECTED → CONFIRMED → IN_PROGRESS → RESOLVED). CONFIRMED(검수확정)가
        "검수완료 + 아직 조치 착수 전"까지 겸한다 — 구 ACTION_PENDING(조치대기)은 CONFIRMED로 흡수되어
        제거됐다(DB V22 마이그레이션).
      enum: [DETECTED, CONFIRMED, IN_PROGRESS, RESOLVED]
    InspectionStatusCode:
      type: string
      enum: [CREATED, UPLOADING, ANALYZING, ANALYZED, REVIEWED, REPORTED]
    InspectionTypeCode:
      type: string
      description: 점검유형(#540 ①) — DB inspection_type PG enum(V7 마이그레이션)과 1:1.
      enum: [REGULAR, DETAILED, EMERGENCY]
    InspectionCreateRequest:
      type: object
      additionalProperties: false
      required: [facilityId, inspectionDate, assignedInspectorId]
      properties:
        facilityId: { type: integer, format: int64, minimum: 1 }
        inspectionDate: { type: string, format: date }
        assignedInspectorId:
          type: integer
          format: int64
          minimum: 1
          description: ACTIVE이며 INSPECTOR 또는 ADMIN 역할인 담당자 식별자
...
    InspectionResponse:
      type: object
      additionalProperties: false
      required:
        [id, facilityId, createdBy, assignedInspectorId, roundNo, inspectionDate, type, status, createdAt]
      properties:
        id: { type: integer, format: int64 }
        facilityId: { type: integer, format: int64 }
        createdBy: { type: integer, format: int64 }
        assignedInspectorId: { type: integer, format: int64 }
        roundNo: { type: integer, minimum: 1 }
        inspectionDate: { type: string, format: date }
        type:
          $ref: "#/components/schemas/InspectionTypeCode"
          description: >
            점검유형(#540 ①) — 엔티티/DB(V7 마이그레이션)에는 이미 있었으나 응답 DTO에 노출되지 않던 갭을
            소급 반영(PR#1032 사람검수 P2-3). 현재 POST /api/inspections 요청 바디에는 이 필드가 없어
            항상 REGULAR로 생성된다(#540 ①은 응답 노출만 범위 — 생성 시 유형 지정은 후속 범위).
        status: { $ref: "#/components/schemas/InspectionStatusCode" }
        createdAt: { type: string, format: date-time }
    InspectionApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/InspectionResponse" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
    InspectionListItemResponse:
      type: object
      additionalProperties: false
      description: >
        GET /api/inspections 목록 항목(#725/HAJA-393, PR #891로 실구현 완료·PR#1032 사람검수 P2-3로
        소급 계약화). assignedInspectorId는 점검 회차 생성 시 배정된 담당자이며, gradeDistribution은
        A~E 5개 키를 항상 포함(값 없는 등급은 0, #893/HAJA-458).
      required:
        [id, facilityId, facilityName, roundNo, inspectionDate, type, assignedInspectorId, assigneeName,
         status, defectCount, gradeDistribution]
      properties:
        id: { type: integer, format: int64 }
        facilityId: { type: integer, format: int64 }
        facilityName: { type: string }
        roundNo: { type: integer, minimum: 1 }
        inspectionDate: { type: string, format: date }
        type: { $ref: "#/components/schemas/InspectionTypeCode" }
        assignedInspectorId: { type: integer, format: int64 }
        assigneeName: { type: string, description: 배정 담당자를 찾을 수 없으면 "-" }
        status: { $ref: "#/components/schemas/InspectionStatusCode" }
        defectCount: { type: integer, format: int64, minimum: 0 }
        gradeDistribution:
          type: object
          description: 등급(A~E)별 하자 건수 — 5개 키 항상 포함
          additionalProperties: { type: integer, format: int64, minimum: 0 }
          example: { A: 0, B: 2, C: 1, D: 0, E: 0 }
    InspectionListItemPage:
      type: object
      description: >
        global.common.PageResponse<InspectionListItemResponse> 매핑(content/page/totalElements 3필드 —
        size/totalPages 는 이 envelope 에 없음, CounselTicketSummaryPage 와 동일 패턴).
      required: [content, page, totalElements]
      properties:
        content:
          type: array
          items: { $ref: "#/components/schemas/InspectionListItemResponse" }
        page: { type: integer, minimum: 0, example: 0 }
        totalElements: { type: integer, format: int64, minimum: 0, example: 12 }
    InspectionListItemPageApiResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean }
        data:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/InspectionListItemPage" }
        error:
          nullable: true
          oneOf:
            - { $ref: "#/components/schemas/ApiError" }
...